@@ -8,11 +8,11 @@ ext {
 
     android = [
             applicationId    : "com.yanzhenjie.album.sample",
-            compileSdkVersion: 28,
-            buildToolsVersion: "28.0.3",
+            compileSdkVersion: 29,
+            buildToolsVersion: "29.0.3",
 
             minSdkVersion    : 14,
-            targetSdkVersion : 28,
+            targetSdkVersion : 29,
 
             versionCode      : 27,
             versionName      : "2.1.3",
@@ -46,10 +46,10 @@ ext {
     dependencies = [
             mediaScanner: 'com.yanzhenjie:mediascanner:1.0.3',
 
-            design      : 'com.android.support:design:28.0.0',
-            appCompat   : 'com.android.support:appcompat-v7:28.0.0',
-            recyclerView: 'com.android.support:recyclerview-v7:28.0.0',
-            cardView    : 'com.android.support:cardview-v7:28.0.0',
+            design      : 'com.google.android.material:material:1.0.0',
+            appCompat   : 'androidx.appcompat:appcompat:1.0.0',
+            recyclerView: 'androidx.recyclerview:recyclerview:1.0.0',
+            cardView    : 'androidx.cardview:cardview:1.0.0',
 
             album       : 'com.yanzhenjie:album:2.1.3',
             glide       : 'com.github.bumptech.glide:glide:3.8.0'
