@@ -16,8 +16,9 @@
 package com.yanzhenjie.album;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import android.net.Uri;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.ImageView;
 
@@ -33,7 +34,7 @@ public interface AlbumLoader {
         }
 
         @Override
-        public void load(ImageView imageView, String url) {
+        public void load(ImageView imageView, Uri url) {
         }
         
         @NonNull
@@ -44,7 +45,7 @@ public interface AlbumLoader {
         
         @NonNull
         @Override
-        public View getPreviewView(Context context, String url, View.OnClickListener onClickListener, View.OnClickListener longClickListener) {
+        public View getPreviewView(Context context, Uri uri, View.OnClickListener onClickListener, View.OnClickListener longClickListener) {
             return new ImageView(context);
         }
         
@@ -62,14 +63,14 @@ public interface AlbumLoader {
      * Load thumbnails of pictures or videos, either local file or remote file.
      *
      * @param imageView {@link ImageView}.
-     * @param url       The url of the file, local path or remote path.
+     * @param uri       The uri of the file, local path or remote path.
      */
-    void load(ImageView imageView, String url);
+    void load(ImageView imageView, Uri uri);
     
     @NonNull
     View getPreviewView(Context context, AlbumFile albumFile, @Nullable View.OnClickListener onClickListener, @Nullable View.OnClickListener longClickListener);
     
     @NonNull
-    View getPreviewView(Context context, String url, @Nullable View.OnClickListener onClickListener, @Nullable View.OnClickListener longClickListener);
+    View getPreviewView(Context context, Uri uri, @Nullable View.OnClickListener onClickListener, @Nullable View.OnClickListener longClickListener);
     
 }
