@@ -16,7 +16,7 @@
 package com.yanzhenjie.album.api;
 
 import android.content.Context;
-import android.support.annotation.IntRange;
+import androidx.annotation.IntRange;
 
 import com.yanzhenjie.album.Filter;
 
