@@ -17,7 +17,7 @@ package com.yanzhenjie.album.api;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.IntRange;
+import androidx.annotation.IntRange;
 
 import com.yanzhenjie.album.Album;
 import com.yanzhenjie.album.app.camera.CameraActivity;
