@@ -19,10 +19,10 @@ import android.content.Context;
 import android.content.res.ColorStateList;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.ColorInt;
-import android.support.annotation.IntDef;
-import android.support.annotation.StringRes;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.ColorInt;
+import androidx.annotation.IntDef;
+import androidx.annotation.StringRes;
+import androidx.core.content.ContextCompat;
 import android.text.TextUtils;
 
 import com.yanzhenjie.album.R;
