@@ -19,11 +19,12 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewCompat;
-import android.support.v4.view.ViewPager;
-import android.support.v7.widget.AppCompatCheckBox;
+import android.net.Uri;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.view.ViewCompat;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.widget.AppCompatCheckBox;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -115,8 +116,8 @@ public class GalleryView<Data> extends Contract.GalleryView<Data> implements Vie
             @NonNull
             @Override
             protected View getView(Context context, Data item, int position, @Nullable View.OnClickListener onClickListener, @Nullable View.OnClickListener longClickListener) {
-                if (item instanceof String) {
-                    return Album.getAlbumConfig().getAlbumLoader().getPreviewView(context, (String) item, onClickListener, longClickListener);
+                if (item instanceof Uri) {
+                    return Album.getAlbumConfig().getAlbumLoader().getPreviewView(context, (Uri) item, onClickListener, longClickListener);
                 } else if (item instanceof AlbumFile) {
                     return Album.getAlbumConfig().getAlbumLoader().getPreviewView(context, (AlbumFile) item, onClickListener, longClickListener);
                 } else {
