@@ -16,7 +16,7 @@
 package com.yanzhenjie.album.provider;
 
 import android.content.Context;
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 /**
  * <p>For external access to files.</p>
