@@ -15,7 +15,7 @@
  */
 package com.yanzhenjie.album.widget.divider;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * <p>Divider of {@code RecyclerView}, you can get the width and height of the line.</p>
