@@ -19,9 +19,6 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.PopupMenu;
 import android.text.TextUtils;
 import android.view.MenuItem;
 import android.view.View;
@@ -29,6 +26,7 @@ import android.widget.CompoundButton;
 
 import com.yanzhenjie.album.Action;
 import com.yanzhenjie.album.Album;
+import com.yanzhenjie.album.AlbumCameraFile;
 import com.yanzhenjie.album.AlbumFile;
 import com.yanzhenjie.album.AlbumFolder;
 import com.yanzhenjie.album.Filter;
@@ -46,57 +44,60 @@ import com.yanzhenjie.album.util.AlbumUtils;
 import com.yanzhenjie.album.widget.LoadingDialog;
 import com.yanzhenjie.mediascanner.MediaScanner;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.widget.PopupMenu;
 
 /**
  * <p>Responsible for controlling the album data and the overall logic.</p>
  * Created by Yan Zhenjie on 2016/10/17.
  */
 public class AlbumActivity extends BaseActivity implements
-        Contract.AlbumPresenter,
-        MediaReadTask.Callback,
-        GalleryActivity.Callback,
-        PathConvertTask.Callback,
-        ThumbnailBuildTask.Callback {
-
+    Contract.AlbumPresenter,
+    MediaReadTask.Callback,
+    GalleryActivity.Callback,
+    PathConvertTask.Callback,
+    ThumbnailBuildTask.Callback {
+    
     private static final int CODE_ACTIVITY_NULL = 1;
     private static final int CODE_PERMISSION_STORAGE = 1;
-
+    
     public static Filter<Long> sSizeFilter;
     public static Filter<String> sMimeFilter;
     public static Filter<Long> sDurationFilter;
-
+    
     public static Action<ArrayList<AlbumFile>> sResult;
     public static Action<String> sCancel;
-
+    
     private List<AlbumFolder> mAlbumFolders;
     private int mCurrentFolder;
-
+    
     private Widget mWidget;
     private int mFunction;
     private int mChoiceMode;
     private int mColumnCount;
     private boolean mHasCamera;
     private int mLimitCount;
-
+    
     private int mQuality;
     private long mLimitDuration;
     private long mLimitBytes;
-
+    
     private boolean mFilterVisibility;
-
+    
     private ArrayList<AlbumFile> mCheckedList;
     private MediaScanner mMediaScanner;
-
+    
     private Contract.AlbumView mView;
     private FolderDialog mFolderDialog;
     private PopupMenu mCameraPopupMenu;
     private LoadingDialog mLoadingDialog;
-
+    
     private MediaReadTask mMediaReadTask;
-
+    
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -107,10 +108,10 @@ public class AlbumActivity extends BaseActivity implements
         mView.setTitle(mWidget.getTitle());
         mView.setCompleteDisplay(false);
         mView.setLoadingDisplay(true);
-
+        
         requestPermission(PERMISSION_STORAGE, CODE_PERMISSION_STORAGE);
     }
-
+    
     private void initializeArgument() {
         Bundle argument = getIntent().getExtras();
         assert argument != null;
@@ -125,7 +126,7 @@ public class AlbumActivity extends BaseActivity implements
         mLimitBytes = argument.getLong(Album.KEY_INPUT_CAMERA_BYTES);
         mFilterVisibility = argument.getBoolean(Album.KEY_INPUT_FILTER_VISIBILITY);
     }
-
+    
     /**
      * Use different layouts depending on the style.
      *
@@ -144,14 +145,16 @@ public class AlbumActivity extends BaseActivity implements
             }
         }
     }
-
+    
     @Override
     public void onConfigurationChanged(Configuration newConfig) {
         super.onConfigurationChanged(newConfig);
         mView.onConfigurationChanged(newConfig);
-        if (mFolderDialog != null && !mFolderDialog.isShowing()) mFolderDialog = null;
+        if (mFolderDialog != null && !mFolderDialog.isShowing()) {
+            mFolderDialog = null;
+        }
     }
-
+    
     @Override
     protected void onPermissionGranted(int code) {
         ArrayList<AlbumFile> checkedList = getIntent().getParcelableArrayListExtra(Album.KEY_INPUT_CHECKED_LIST);
@@ -159,22 +162,22 @@ public class AlbumActivity extends BaseActivity implements
         mMediaReadTask = new MediaReadTask(mFunction, checkedList, mediaReader, this);
         mMediaReadTask.execute();
     }
-
+    
     @Override
     protected void onPermissionDenied(int code) {
         new AlertDialog.Builder(this)
-                .setCancelable(false)
-                .setTitle(R.string.album_title_permission_failed)
-                .setMessage(R.string.album_permission_storage_failed_hint)
-                .setPositiveButton(R.string.album_ok, new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        callbackCancel();
-                    }
-                })
-                .show();
+            .setCancelable(false)
+            .setTitle(R.string.album_title_permission_failed)
+            .setMessage(R.string.album_permission_storage_failed_hint)
+            .setPositiveButton(R.string.album_ok, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int which) {
+                    callbackCancel();
+                }
+            })
+            .show();
     }
-
+    
     @Override
     public void onScanCallback(ArrayList<AlbumFolder> albumFolders, ArrayList<AlbumFile> checkedFiles) {
         mMediaReadTask = null;
@@ -191,11 +194,11 @@ public class AlbumActivity extends BaseActivity implements
                 throw new AssertionError("This should not be the case.");
             }
         }
-
+        
         mView.setLoadingDisplay(false);
         mAlbumFolders = albumFolders;
         mCheckedList = checkedFiles;
-
+        
         if (mAlbumFolders.get(0).getAlbumFiles().isEmpty()) {
             Intent intent = new Intent(this, NullActivity.class);
             intent.putExtras(getIntent());
@@ -207,15 +210,17 @@ public class AlbumActivity extends BaseActivity implements
             mView.setSubTitle(count + "/" + mLimitCount);
         }
     }
-
+    
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         switch (requestCode) {
             case CODE_ACTIVITY_NULL: {
                 if (resultCode == RESULT_OK) {
-                    String imagePath = NullActivity.parsePath(data);
-                    String mimeType = AlbumUtils.getMimeType(imagePath);
-                    if (!TextUtils.isEmpty(mimeType)) mCameraAction.onAction(imagePath);
+                    AlbumCameraFile albumCameraFile = NullActivity.parseAlbumCameraFile(data);
+                    String mimeType = AlbumUtils.getMimeType(albumCameraFile.getPath());
+                    if (!TextUtils.isEmpty(mimeType)) {
+                        mCameraAction.onAction(albumCameraFile);
+                    }
                 } else {
                     callbackCancel();
                 }
@@ -223,7 +228,7 @@ public class AlbumActivity extends BaseActivity implements
             }
         }
     }
-
+    
     @Override
     public void clickFolderSwitch() {
         if (mFolderDialog == null) {
@@ -235,9 +240,11 @@ public class AlbumActivity extends BaseActivity implements
                 }
             });
         }
-        if (!mFolderDialog.isShowing()) mFolderDialog.show();
+        if (!mFolderDialog.isShowing()) {
+            mFolderDialog.show();
+        }
     }
-
+    
     /**
      * Update data source.
      */
@@ -246,7 +253,7 @@ public class AlbumActivity extends BaseActivity implements
         AlbumFolder albumFolder = mAlbumFolders.get(position);
         mView.bindAlbumFolder(albumFolder);
     }
-
+    
     @Override
     public void clickCamera(View v) {
         int hasCheckSize = mCheckedList == null ? 0 : mCheckedList.size();
@@ -306,80 +313,87 @@ public class AlbumActivity extends BaseActivity implements
             }
         }
     }
-
+    
     private void takePicture() {
         String filePath;
-        if (mCurrentFolder == 0) {
-            filePath = AlbumUtils.randomJPGPath();
-        } else {
-            File file = new File(mAlbumFolders.get(mCurrentFolder).getAlbumFiles().get(0).getPath());
-            filePath = AlbumUtils.randomJPGPath(file.getParentFile());
-        }
+        filePath = AlbumUtils.randomJPGPath(this.getApplicationContext());
+        //        if (mCurrentFolder == 0) {
+        //            filePath = AlbumUtils.randomJPGPath(this.getApplicationContext());
+        //        } else {
+        ////            File file = new File(mAlbumFolders.get(mCurrentFolder).getAlbumFiles().get(0).getPath());
+        //            filePath = AlbumUtils.randomJPGPath(this.getApplicationContext());
+        //        }
         Album.camera(this)
-                .image()
-                .filePath(filePath)
-                .onResult(mCameraAction)
-                .start();
+            .image()
+            .filePath(filePath)
+            .onResult(mCameraAction)
+            .start();
     }
-
+    
     private void takeVideo() {
         String filePath;
-        if (mCurrentFolder == 0) {
-            filePath = AlbumUtils.randomMP4Path();
-        } else {
-            File file = new File(mAlbumFolders.get(mCurrentFolder).getAlbumFiles().get(0).getPath());
-            filePath = AlbumUtils.randomMP4Path(file.getParentFile());
-        }
+        //        if (mCurrentFolder == 0) {
+        //            fileUri = AlbumUtils.randomMP4Path(this.getApplicationContext());
+        //        } else {
+        //            File file = new File(mAlbumFolders.get(mCurrentFolder).getAlbumFiles().get(0).getPath());
+        //            fileUri = AlbumUtils.randomMP4Path(this.getApplicationContext(), file.getParentFile());
+        //        }
+        filePath = AlbumUtils.randomMP4Path(this.getApplicationContext());
         Album.camera(this)
-                .video()
-                .filePath(filePath)
-                .quality(mQuality)
-                .limitDuration(mLimitDuration)
-                .limitBytes(mLimitBytes)
-                .onResult(mCameraAction)
-                .start();
+            .video()
+            .filePath(filePath)
+            .quality(mQuality)
+            .limitDuration(mLimitDuration)
+            .limitBytes(mLimitBytes)
+            .onResult(mCameraAction)
+            .start();
     }
-
-    private Action<String> mCameraAction = new Action<String>() {
+    
+    private Action<AlbumCameraFile> mCameraAction = new Action<AlbumCameraFile>() {
         @Override
-        public void onAction(@NonNull String result) {
+        public void onAction(@NonNull AlbumCameraFile result) {
             if (mMediaScanner == null) {
                 mMediaScanner = new MediaScanner(AlbumActivity.this);
             }
-            mMediaScanner.scan(result);
-
+            mMediaScanner.scan(result.getPath());
             PathConversion conversion = new PathConversion(sSizeFilter, sMimeFilter, sDurationFilter);
-            PathConvertTask task = new PathConvertTask(conversion, AlbumActivity.this);
-            task.execute(result);
+            PathConvertTask task = new PathConvertTask(AlbumActivity.this, conversion, AlbumActivity.this);
+            task.execute(result.getPath());
         }
     };
-
+    
     @Override
     public void onConvertStart() {
         showLoadingDialog();
         mLoadingDialog.setMessage(R.string.album_converting);
     }
-
+    
     @Override
     public void onConvertCallback(AlbumFile albumFile) {
         albumFile.setChecked(!albumFile.isDisable());
         if (albumFile.isDisable()) {
-            if (mFilterVisibility) addFileToList(albumFile);
-            else mView.toast(getString(R.string.album_take_file_unavailable));
+            if (mFilterVisibility) {
+                addFileToList(albumFile);
+            } else {
+                mView.toast(getString(R.string.album_take_file_unavailable));
+            }
         } else {
             addFileToList(albumFile);
         }
-
+        
         dismissLoadingDialog();
     }
-
+    
     private void addFileToList(AlbumFile albumFile) {
         if (mCurrentFolder != 0) {
             List<AlbumFile> albumFiles = mAlbumFolders.get(0).getAlbumFiles();
-            if (albumFiles.size() > 0) albumFiles.add(0, albumFile);
-            else albumFiles.add(albumFile);
+            if (albumFiles.size() > 0) {
+                albumFiles.add(0, albumFile);
+            } else {
+                albumFiles.add(albumFile);
+            }
         }
-
+        
         AlbumFolder albumFolder = mAlbumFolders.get(mCurrentFolder);
         List<AlbumFile> albumFiles = albumFolder.getAlbumFiles();
         if (albumFiles.isEmpty()) {
@@ -389,12 +403,12 @@ public class AlbumActivity extends BaseActivity implements
             albumFiles.add(0, albumFile);
             mView.notifyInsertItem(mHasCamera ? 1 : 0);
         }
-
+        
         mCheckedList.add(albumFile);
         int count = mCheckedList.size();
         mView.setCheckedCount(count);
         mView.setSubTitle(count + "/" + mLimitCount);
-
+        
         switch (mChoiceMode) {
             case Album.MODE_SINGLE: {
                 callbackResult();
@@ -409,7 +423,7 @@ public class AlbumActivity extends BaseActivity implements
             }
         }
     }
-
+    
     @Override
     public void tryCheckItem(CompoundButton button, int position) {
         AlbumFile albumFile = mAlbumFolders.get(mCurrentFolder).getAlbumFiles().get(position);
@@ -446,23 +460,23 @@ public class AlbumActivity extends BaseActivity implements
             setCheckedCount();
         }
     }
-
+    
     private void setCheckedCount() {
         int count = mCheckedList.size();
         mView.setCheckedCount(count);
         mView.setSubTitle(count + "/" + mLimitCount);
     }
-
+    
     @Override
     public void tryPreviewItem(int position) {
         switch (mChoiceMode) {
             case Album.MODE_SINGLE: {
                 AlbumFile albumFile = mAlbumFolders.get(mCurrentFolder).getAlbumFiles().get(position);
-//                albumFile.setChecked(true);
-//                mView.notifyItem(position);
+                //                albumFile.setChecked(true);
+                //                mView.notifyItem(position);
                 mCheckedList.add(albumFile);
                 setCheckedCount();
-
+                
                 callbackResult();
                 break;
             }
@@ -481,7 +495,7 @@ public class AlbumActivity extends BaseActivity implements
             }
         }
     }
-
+    
     @Override
     public void tryPreviewChecked() {
         if (mCheckedList.size() > 0) {
@@ -494,27 +508,31 @@ public class AlbumActivity extends BaseActivity implements
             startActivity(intent);
         }
     }
-
+    
     @Override
     public void onPreviewComplete() {
         callbackResult();
     }
-
+    
     @Override
     public void onPreviewChanged(AlbumFile albumFile) {
         ArrayList<AlbumFile> albumFiles = mAlbumFolders.get(mCurrentFolder).getAlbumFiles();
         int position = albumFiles.indexOf(albumFile);
         int notifyPosition = mHasCamera ? position + 1 : position;
         mView.notifyItem(notifyPosition);
-
+        
         if (albumFile.isChecked()) {
-            if (!mCheckedList.contains(albumFile)) mCheckedList.add(albumFile);
+            if (!mCheckedList.contains(albumFile)) {
+                mCheckedList.add(albumFile);
+            }
         } else {
-            if (mCheckedList.contains(albumFile)) mCheckedList.remove(albumFile);
+            if (mCheckedList.contains(albumFile)) {
+                mCheckedList.remove(albumFile);
+            }
         }
         setCheckedCount();
     }
-
+    
     @Override
     public void complete() {
         if (mCheckedList.isEmpty()) {
@@ -541,42 +559,48 @@ public class AlbumActivity extends BaseActivity implements
             callbackResult();
         }
     }
-
+    
     @Override
     public void onBackPressed() {
-        if (mMediaReadTask != null) mMediaReadTask.cancel(true);
+        if (mMediaReadTask != null) {
+            mMediaReadTask.cancel(true);
+        }
         callbackCancel();
     }
-
+    
     /**
      * Callback result action.
      */
     private void callbackResult() {
-        ThumbnailBuildTask task = new ThumbnailBuildTask(this, mCheckedList, this);
+        ThumbnailBuildTask task = new ThumbnailBuildTask(this.getApplicationContext(), mCheckedList, this);
         task.execute();
     }
-
+    
     @Override
     public void onThumbnailStart() {
         showLoadingDialog();
         mLoadingDialog.setMessage(R.string.album_thumbnail);
     }
-
+    
     @Override
     public void onThumbnailCallback(ArrayList<AlbumFile> albumFiles) {
-        if (sResult != null) sResult.onAction(albumFiles);
+        if (sResult != null) {
+            sResult.onAction(albumFiles);
+        }
         dismissLoadingDialog();
         finish();
     }
-
+    
     /**
      * Callback cancel action.
      */
     private void callbackCancel() {
-        if (sCancel != null) sCancel.onAction("User canceled.");
+        if (sCancel != null) {
+            sCancel.onAction("User canceled.");
+        }
         finish();
     }
-
+    
     /**
      * Display loading dialog.
      */
@@ -589,7 +613,7 @@ public class AlbumActivity extends BaseActivity implements
             mLoadingDialog.show();
         }
     }
-
+    
     /**
      * Dismiss loading dialog.
      */
@@ -598,7 +622,7 @@ public class AlbumActivity extends BaseActivity implements
             mLoadingDialog.dismiss();
         }
     }
-
+    
     @Override
     public void finish() {
         sSizeFilter = null;
