@@ -26,13 +26,13 @@ import java.util.ArrayList;
  * Created by YanZhenjie on 2017/10/15.
  */
 public class ThumbnailBuildTask extends AsyncTask<Void, Void, ArrayList<AlbumFile>> {
-
+    
     public interface Callback {
         /**
          * The task begins.
          */
         void onThumbnailStart();
-
+        
         /**
          * Callback results.
          *
@@ -40,36 +40,41 @@ public class ThumbnailBuildTask extends AsyncTask<Void, Void, ArrayList<AlbumFil
          */
         void onThumbnailCallback(ArrayList<AlbumFile> albumFiles);
     }
-
+    
     private ArrayList<AlbumFile> mAlbumFiles;
     private Callback mCallback;
-
+    
     private ThumbnailBuilder mThumbnailBuilder;
-
+    private Context mContext;
+    
     public ThumbnailBuildTask(Context context, ArrayList<AlbumFile> albumFiles, Callback callback) {
         this.mAlbumFiles = albumFiles;
+        if (mAlbumFiles == null) {
+            mAlbumFiles = new ArrayList<>();
+        }
         this.mCallback = callback;
-        this.mThumbnailBuilder = new ThumbnailBuilder(context);
+        this.mContext = context;
+        this.mThumbnailBuilder = new ThumbnailBuilder(mContext);
     }
-
+    
     @Override
     protected void onPreExecute() {
         mCallback.onThumbnailStart();
     }
-
+    
     @Override
     protected ArrayList<AlbumFile> doInBackground(Void... params) {
         for (AlbumFile albumFile : mAlbumFiles) {
             int mediaType = albumFile.getMediaType();
             if (mediaType == AlbumFile.TYPE_IMAGE) {
-                albumFile.setThumbPath(mThumbnailBuilder.createThumbnailForImage(albumFile.getPath()));
+                albumFile.setThumbPath(mThumbnailBuilder.createThumbnailForImage(mContext, albumFile.getUri(), albumFile.getMimeType()));
             } else if (mediaType == AlbumFile.TYPE_VIDEO) {
-                albumFile.setThumbPath(mThumbnailBuilder.createThumbnailForVideo(albumFile.getPath()));
+                albumFile.setThumbPath(mThumbnailBuilder.createThumbnailForVideo(mContext, albumFile.getUri()));
             }
         }
         return mAlbumFiles;
     }
-
+    
     @Override
     protected void onPostExecute(ArrayList<AlbumFile> albumFiles) {
         mCallback.onThumbnailCallback(albumFiles);
