@@ -15,7 +15,7 @@
  */
 package com.yanzhenjie.album.mvp;
 
-import android.arch.lifecycle.LifecycleOwner;
+import androidx.lifecycle.LifecycleOwner;
 
 /**
  * <p>Presenter of MVP.</p>
